requirements.txt
Flask
flask-socketio
flask-migrate
flask-sqlalchemy
markdown
pymdown-extensions
omegaconf
numpy<2
scipy
torch
torchaudio
send2trash
Pillow
requests
transformers
huggingface-hub
tqdm
opencv-python
imageio
scikit-learn
datasets
pydub
python-dateutil
tinytag
unidecode
mutagen
peft
trl
pandas
sentencepiece
einops 
sentence-transformers
Flask-HTTPAuth
protobuf
moviepy==1.0.3
watchdog